Update glslang using temporary branch, fix LtRtProcessing

+++ lib/audiodev/LtRtProcessing.cpp
@@ -18,7 +18,7 @@ constexpr T ClampFull(float in) {
 
     if (in < MIN)
       return MIN;
-    else if (in > MAX)
+    else if (in > static_cast<float>(MAX))
       return MAX;
     else
       return in;
